+++ package.json
@@ -6,35 +6,35 @@
   "dependencies": {
     "bulma": "^0.9.0",
     "classnames": "^2.3.1",
-    "gatsby": "^2.20.35",
+    "gatsby": "^3.6.1",
     "gatsby-background-image": "^1.5.3",
-    "gatsby-image": "^2.3.5",
+    "gatsby-image": "^3.6.0",
     "gatsby-plugin-copy-files": "^1.0.4",
-    "gatsby-plugin-netlify": "^2.2.4",
-    "gatsby-plugin-netlify-cms": "^4.2.5",
-    "gatsby-plugin-purgecss": "^5.0.0",
-    "gatsby-plugin-react-helmet": "^3.2.5",
-    "gatsby-plugin-sass": "^2.2.4",
-    "gatsby-plugin-sharp": "^2.5.7",
+    "gatsby-plugin-netlify": "^3.6.1",
+    "gatsby-plugin-netlify-cms": "^5.6.0",
+    "gatsby-plugin-purgecss": "^6.0.2",
+    "gatsby-plugin-react-helmet": "^4.6.0",
+    "gatsby-plugin-sass": "^4.6.0",
+    "gatsby-plugin-sharp": "^3.6.0",
     "gatsby-plugin-web-font-loader": "^1.0.4",
-    "gatsby-remark-copy-linked-files": "^2.2.4",
-    "gatsby-remark-images": "^3.2.6",
+    "gatsby-remark-copy-linked-files": "^4.3.0",
+    "gatsby-remark-images": "^5.3.0",
     "gatsby-remark-relative-images": "^0.3.0",
-    "gatsby-source-filesystem": "^2.2.5",
-    "gatsby-transformer-remark": "^2.7.5",
-    "gatsby-transformer-sharp": "^2.4.7",
+    "gatsby-source-filesystem": "^3.6.0",
+    "gatsby-transformer-remark": "^4.3.0",
+    "gatsby-transformer-sharp": "^3.6.0",
     "lodash": "^4.17.15",
     "lodash-webpack-plugin": "^0.11.4",
-    "netlify-cms-app": "^2.14.46",
-    "netlify-cms-locales": "^1.30.0",
+    "netlify-cms-app": "^2.15.11",
+    "netlify-cms-locales": "^1.30.1",
     "netlify-cms-media-library-cloudinary": "^1.3.10",
     "netlify-cms-media-library-uploadcare": "^0.5.10",
-    "node-sass": "^4.14.0",
+    "node-sass": "^6.0.0",
     "prop-types": "^15.6.0",
-    "react": "^16.8.4",
-    "react-dom": "^16.8.4",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
     "react-helmet": "^6.0.0",
-    "uuid": "^7.0.0"
+    "uuid": "^8.3.2"
   },
   "keywords": [
     "gatsby"
@@ -50,6 +50,6 @@
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "devDependencies": {
-    "prettier": "^2.0.5"
+    "prettier": "^2.3.0"
   }
 }
