+++ src/views/news-post.js
