+++ src/views/private-policy-page.js
